Fix crush depth activating even when set to Normal.
